requirements.txt
torch==2.5.1+rocm6.2
torchvision==0.20.1+rocm6.2
transformers>=4.45.0
peft>=0.13.0
accelerate>=1.0.0
bitsandbytes>=0.44.0
datasets>=2.0.0
networkx>=3.0
chromadb>=0.5.0
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
tqdm>=4.66.0
rich>=13.0.0
matplotlib>=3.9.0
